revert: convert `upper_imm` to unsigned int32"
This reverts commit 81702424d32fd6aa55edbbaf8e68be173e3600f5.

Turns out this is needed to keep RV64 tests accurate.

+++ core/core/sim/cores/cpu/rv32/rv_i_instructions.cpp
@@ -35,10 +35,10 @@ template <u32 (*Op)(u32, u32)> void reg_imm(RV32CPU *self, riscv::InstructionI i
 } // namespace
 
 // Upper immediate
-void handle_lui(RV32CPU *self, riscv::InstructionU u) { self->write_register((XReg)u.rd, u.upper_imm()); }
+void handle_lui(RV32CPU *self, riscv::InstructionU u) { self->write_register((XReg)u.rd, u32(u.upper_imm())); }
 // Relative to this instruction's own address. Needs PC prior to execution.
 void handle_auipc(RV32CPU *self, riscv::InstructionU u) {
-  self->write_register((XReg)u.rd, self->read_initial_pc() + u.upper_imm());
+  self->write_register((XReg)u.rd, self->read_initial_pc() + u32(u.upper_imm()));
 }
 
 // Unconditional jumps
